refactor(cli): move transcript build/render to tui_detail_transcript.go (PRD #1009 M3)
